+++ src/flows/CostCalculator/index.ts
@@ -1,3 +1,3 @@
 export { CostCalculator } from './CostCalculator';
 export { CostCalculatorResults } from './Results/CostCalculatorResults';
-export { useCostCalculator, useExportAsPdf } from './hooks';
+export { useCostCalculator, useCostCalculatorEstimationPdf } from './hooks';
